requirements.txt
Flask==2.0.1
requests==2.26.0
tensorflow==2.7.0
google-cloud-storage==2.17.0
Werkzeug==2.2.2
protobuf==3.20.*
Pillow==8.4.0
